+++ packages/eas-cli/src/commands/upload.ts
@@ -11,7 +11,10 @@ import { v4 as uuidv4 } from 'uuid';
 import { getBuildLogsUrl } from '../build/utils/url';
 import EasCommand from '../commandUtils/EasCommand';
 import { ExpoGraphqlClient } from '../commandUtils/context/contextUtils/createGraphqlClient';
-import { EasNonInteractiveAndJsonFlags } from '../commandUtils/flags';
+import {
+  EasNonInteractiveAndJsonFlags,
+  resolveNonInteractiveAndJsonFlags,
+} from '../commandUtils/flags';
 import {
   BuildMetadataInput,
   DistributionType,
@@ -55,12 +58,8 @@ export default class BuildUpload extends EasCommand {
 
   async runAsync(): Promise<void> {
     const { flags } = await this.parse(BuildUpload);
-    const {
-      'build-path': buildPath,
-      fingerprint: manualFingerprintHash,
-      json: jsonFlag,
-      'non-interactive': nonInteractive,
-    } = flags;
+    const { 'build-path': buildPath, fingerprint: manualFingerprintHash } = flags;
+    const { json: jsonFlag, nonInteractive } = resolveNonInteractiveAndJsonFlags(flags);
     const {
       projectId,
       loggedIn: { graphqlClient },